requirements.txt
anthropic>=0.40.0
voyageai>=0.3.2
openai>=1.50.0
google-genai>=1.0.0
python-dotenv>=1.0.0
